+++ kotlin-collections-projeto-inicial/src/br/com/alura/main.kt
@@ -1,27 +1,16 @@
 package br.com.alura
 
 fun main() {
-    val assistiramCursoAndroid: Set<String> = setOf("Alex", "Fran", "Gui", "Maria")
-    val assistiramCursoKotlin: Set<String> = setOf("Alex", "Paulo", "Maria")
-    val assistiramAmbos = mutableSetOf<String>()
-    assistiramAmbos.addAll(assistiramCursoAndroid)
-    assistiramAmbos.addAll(assistiramCursoKotlin)
-    assistiramAmbos.add("Ana")
-    println(assistiramAmbos)
+    val pedidos = mapOf(Pair(1, 20.0), Pair(2, 34.0), 3 to 50.0)
+    println(pedidos)
 
-    //Função Union
-    println(assistiramCursoAndroid + assistiramCursoKotlin)
-    println(assistiramCursoAndroid union assistiramCursoKotlin)
+    val pedido = pedidos[1]
+    pedido?.let {
+        println("pedido $it")
+    }
 
-    //Função Subtract
-    println(assistiramCursoKotlin - assistiramCursoAndroid)
-    println(assistiramCursoKotlin subtract assistiramCursoAndroid)
-
-    //Função Intersect
-    println(assistiramCursoKotlin intersect assistiramCursoAndroid)
-
-    val assistiramList = assistiramAmbos.toMutableList()
-    assistiramList.add("Alex")
-    println(assistiramList)
-    println(assistiramList.toSet())
+    for (pedido: Map.Entry<Int, Double> in pedidos) {
+        println("Número do pedido: ${pedido.key}")
+        println("Valor do pedido: ${pedido.value}")
+    }
 }
